gps_srv not quite updated
